# shared/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Review Anything API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/config/": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_config__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Config Api Config  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Create Or Update Config",
        "operationId": "create_or_update_config_api_config__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "config"
        ],
        "summary": "Delete Config",
        "description": "删除模型配置",
        "operationId": "delete_config_api_config__delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/config/test": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "Test Config Connection",
        "description": "测试 LLM API 连接是否正常",
        "operationId": "test_config_connection_api_config_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigCreate": {
        "properties": {
          "api_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "default": "gpt-3.5-turbo"
          },
          "daily_review_count": {
            "type": "integer",
            "title": "Daily Review Count",
            "default": 10
          }
        },
        "type": "object",
        "title": "ConfigCreate"
      },
      "ConfigResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "api_base_url": {
            "type": "string",
            "title": "Api Base Url"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "daily_review_count": {
            "type": "integer",
            "title": "Daily Review Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "api_base_url",
          "model_name",
          "daily_review_count"
        ],
        "title": "ConfigResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}